# Optimize spherical manifold get_new_point into get_new_points