--- a/doc/news/changes/minor/20171201ReneGassmoeller
+++ b/doc/news/changes/minor/20171201ReneGassmoeller
@@ -1,0 +1,5 @@
+Improved: SphericalManifold::get_new_points now computes a lot
+of information outside of the loop over all points in get_new_point,
+which saves a significant amount of time for spherical geometries.
+<br>
+(Rene Gassmoeller, 2017/12/01)