… tests/grid/project_to_object_01.output
… tests/manifold/transfinite_manifold_03.output