This PR is on top of #5529 and related to #5506.

First of all sorry for not coming back to this earlier. @masterleinad I checked #5529, and it indeed improves timings for large computations (when less points are computed in the Newton iteration), but it is still significantly slower than before. I suppose that is because now more instructions are done for each point, even for the simple case (e.g. the check for max_distance). I adapted the idea of @kronbichler in #5506 and moved all operations from `get_new_point` that do not depend on the weights into a new `SphericalManifold::get_new_points` functions. Although it is a pain to get all special cases correct, now I am essentially back to the performance before the introduction of the Newton iteration (while still having the increased accuracy of it). I show some timings below.  This new `get_new_points` functions could also allow further optimization as mentioned here https://github.com/dealii/dealii/issues/5506#issuecomment-345641636

Although it is hard to see, the logic in this PR should stay unchanged compared to #5529, except I now use the maximum distance between the surrounding points as upper bound for the maximum distance between candidate point and surrounding points. This seems reasonable as the candidate point lies in between the surrounding points. With this approximation the maximum distance check is independent of the weights.

Some timings below. All were taken from ASPECT, showing all steps of the assembly of an advection-diffusion system, for a 3D spherical shell (Rinner ~ 0.55 Router) with 4 refinements (393,216 cells, 3,195,010 Dofs for a Q2 element), and a MappingQGeneric(4) on 40 MPI processes.

Version | Time
------------ | -------------
8.5 |  6.3s
9.0 pre #5368   | 6.27s
9.0 pre #5529 | 9.27s
9.0 with #5529 | 7.24s
9.0 with this PR | 6.19s

I guess it is hard to generalize to all refinements and special cases, but for the moment I would be happy with this result. @masterleinad: I would appreciate if you could take a look at the code if I broke something (spherical manifold tests seem to work on my machine). Btw: I did not change anything in the Newton iteration itself, so if the discussion in #5506 leads to improvements in the Newton loop that might speed things up further.