@kronbichler 2017-12-02
@gassmoeller This looks really nice at first sight. Thanks for following up on my idea, it is indeed what I had in mind. I will have a closer look with code-specific comments during the following days.