@masterleinad 2017-12-02
Thanks for doing this!
This reduces the cost for my modified step-10 testcase from `148*10^9` instructions (#5529) to `135*10^9` instructions. So even for rather fine meshes there still is like a 10% gain.

I'll reevaluate if the threshold 10^-2 is still good or can get a bit larger.

Apart from that we need to think on how to proceed with this. In the end, this just supersedes #5529.