@masterleinad 2017-12-02
For me the following tests fail (locally):
```
     2085 - bits/fe_tools_02.debug (Failed)
     2087 - bits/fe_tools_03.debug (Failed)
     2559 - codim_one/mapping_fe_field_01.debug (Failed)
     2560 - codim_one/mapping_fe_field_01.release (Failed)
     4464 - grid/project_to_object_01.debug (Failed)
     4465 - grid/project_to_object_01.release (Failed)
     4486 - grid/torus_01.debug (Failed)
     4487 - grid/torus_01.release (Failed)
     5606 - manifold/chart_manifold_10.debug (Failed)
     5607 - manifold/chart_manifold_10.release (Failed)
     5848 - matrix_free/copy_feevaluation.debug (Failed)
     6014 - matrix_free/matrix_vector_stokes_noflux.debug (Failed)
     6016 - matrix_free/matrix_vector_stokes_notempl.debug (Failed)
```

In particular, the failing `matrix_free` tests report
```
--------------------------------------------------------
 6016: An error occurred in line <2894> of file <../source/fe/mapping_q_generic.cc> in function
 6016:     virtual CellSimilarity::Similarity dealii::MappingQGeneric<3, 3>::fill_fe_values(const typename Triangulation<dim, spacedim>::cell_iterator &, const CellSimilarity::Similarity,      const Quadrature<dim> &, const typename Mapping<dim, spacedim>::InternalDataBase &, internal::FEValues::MappingRelatedData<dim, spacedim> &) const [dim = 3, spacedim = 3]
 6016: The violated condition was:
 6016:     det > 1e-12*Utilities::fixed_power<dim>(cell->diameter()/ std::sqrt(double(dim)))
 6016: Additional information:
 6016:     The image of the mapping applied to cell with center [-0.364232 -0.364232 0.179005] is distorted. The cell geometry or the mapping are invalid, giving a non-positive volume          fraction of -0.00664757 in quadrature point 13.
 6016: --------------------------------------------------------
```