@kronbichler 2017-12-02
@masterleinad I have not looked in detail, but there is a fix in d50e3419 in #5386 that will in turn also need #5564, so I don't think we should bother about those three tests here. (Unless it means that something fundamental in the spherical manifold is broken, of course.)