@masterleinad 2017-12-02
@kronbichler Yes, applying https://github.com/dealii/dealii/commit/d50e34194b478019cd69afadcd54ea772e9f19c5 fixes these failing tests. Would you mind to cherry-pick this into a separate PR?