@masterleinad 2017-12-02
> I have one larger request/idea: After merging the weights "horizonally" (between different points), shouldn't we also be able to merge weights "vertically"? This would reduce the size of the loop in line 487 that does the get_new_point with the Newton iteration in the current version, but we would need to only return the unit point and add the different radii outside of that loop.

We should try that.