@masterleinad 2017-12-03
With the new check for the maximum distance between `directions` the minimum distance that still requires a second Newton iterate is consistently around 0.113 from the second refinement on in my testcase.
The minimum distance for calls converging in the first Newton iterates are:
```
ref 1: 0.222
ref 2: 0.0507
ref 3: 0.0116
ref 4: 0.0273
```
and behaves as `ref^-2` resp. `h^2`. Hence, I would suggest to increase this tolerance a bit to 2.e-2 so that we benefit already in the third refinement. Since we are a factor 5 away from the minimum value in the second iterate, this should still be fine.