@gassmoeller 2017-12-03
I think I addressed most comments except for the `tolerance` class member (because it is now used less often in the class, but instead in a non-member helper function that I need to do partial specialization).
The horizontal merge indeed significantly improves the speed of the Newton iteration (or rather calls the Newton iteration 5 times less often for a hyper shell with 0 refinement and MappingQ4), while the speed for the linear guess remains more or less constant. I suppose this is the best we can do for now. Please take a look if you approve of my specialization / function-call scheme.