@masterleinad 2017-12-04
Hmm... I don't think that we can ignore the failing tests (by updating the output) at least for `mapping_q_mixed_manifolds_02` the diff
```
  DEAL::Volume 2D mapping degree 1: 0.156929 error: 0.00501399
 5680:   DEAL::Volume 2D mapping degree 2: 0.156152 error: 3.91474e-05
 5680: ! DEAL::Volume 2D mapping degree 3: 0.156146 error: 1.53930e-07
 5680: ! DEAL::Volume 2D mapping degree 4: 0.156146 error: 3.54084e-10
 5680: ! DEAL::Volume 2D mapping degree 5: 0.156146 error: 5.29174e-13
 5680: ! DEAL::Volume 2D mapping degree 6: 0.156146 error: -7.11016e-16
 5680:   DEAL::Volume 3D mapping degree 1: 0.0643409 error: 0.00501399
 5680:   DEAL::Volume 3D mapping degree 2: 0.0640224 error: 3.91474e-05
 5680:   DEAL::Volume 3D mapping degree 3: 0.0640199 error: 1.53930e-07
 5680: --- 1,10 ----
 5680:
 5680:   DEAL::Volume 2D mapping degree 1: 0.156929 error: 0.00501399
 5680:   DEAL::Volume 2D mapping degree 2: 0.156152 error: 3.91474e-05
 5680: ! DEAL::Volume 2D mapping degree 3: 0.156144 error: -1.09871e-05
 5680: ! DEAL::Volume 2D mapping degree 4: 0.156146 error: -1.10148e-07
 5680: ! DEAL::Volume 2D mapping degree 5: 0.156146 error: -1.66407e-08
 5680: ! DEAL::Volume 2D mapping degree 6: 0.156146 error: -1.69919e-10
 5680:   DEAL::Volume 3D mapping degree 1: 0.0643409 error: 0.00501399
 5680:   DEAL::Volume 3D mapping degree 2: 0.0640224 error: 3.91474e-05
 5680:   DEAL::Volume 3D mapping degree 3: 0.0640199 error: 1.53930e-07
```
shows a serious deterioration. Varying the tolerance for `max_distance` does not help here.