@masterleinad 2017-12-04
Removing the last commit ("Merge horizontally") helps.