@kronbichler 2017-12-04
@masterleinad The failure is in 2D, not in 3D, where we should not use `SphericalManifold`. Apart from the horizontal merging (which we still must verify carefully), I think we should fix the 2D issue on a more fundamental level by creating a specialization of `SphericalManifold` for `spacedim=2` that derives from `PolarManifold` (rather than using a member). That should be much cleaner.