@masterleinad 2017-12-05
@gassmoeller Would you mind to rebase? Since #5564 is merged the number of failing tests should be much smaller.