@masterleinad 2017-12-13
@gassmoeller Can you please rebase? After #5386, there might be some changes in the test output. I would really like to get this merged soon.